tas/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pocus"
version = "0.1.0"
description = "POCUS - ポーカー店舗情報集約システム"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # HTTP client
    "httpx>=0.27.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Headless browser (SPA対応)
    "playwright>=1.44.0",
    # HTML parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    # Configuration
    "pydantic-settings>=2.2.0",
    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",
    # Google Sheets
    "google-api-python-client>=2.128.0",
    "google-auth>=2.29.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "ujson>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.21.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
pocus = "tas.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["tas"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
